+++ publish-cli/src/main/java/com/ericsson/eiffel/remrem/publish/cli/CliOptions.java
@@ -15,6 +15,7 @@
 package com.ericsson.eiffel.remrem.publish.cli;
 
 import java.util.ArrayList;
+import java.util.Map;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -27,6 +28,7 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.lang3.ArrayUtils;
 
 import com.ericsson.eiffel.remrem.publish.config.PropertiesConfig;
+import com.ericsson.eiffel.remrem.shared.VersionService;
 
 public class CliOptions {
 	
@@ -77,6 +79,7 @@ public class CliOptions {
         options.addOption("mp", "messaging_protocol", true, "name of messaging protocol to be used, e.g. eiffel3, eiffelsemantics, default is eiffelsemantics");
         options.addOption("domain", "domainId", true, "identifies the domain that produces the event");
         options.addOption("ud", "user_domain_suffix", true, "user domain suffix");
+        options.addOption("v", "lists the versions of publish and all loaded protocols");
         contentGroup = createContentGroup();
         options.addOptionGroup(contentGroup);
     }
@@ -124,7 +127,9 @@ public class CliOptions {
         if (commandLine.hasOption("h")) {
     	    System.out.println("You passed help flag.");
     	    help(0);
-        } else {
+        } else if (commandLine.hasOption("v")) {
+            printVersions();
+        }else {
             checkRequiredOptions();
         }
     }
@@ -263,5 +268,29 @@ public class CliOptions {
         addErrorCode(errorCode);
       else
         System.exit(errorCode);
+    }
+    
+    /**
+     * Lists the versions of publish and all loaded protocols  
+     */
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    private static void printVersions() {
+        Map versions = new VersionService().getMessagingVersions();
+        Map<String, String> endpointVersions = (Map<String, String>) versions.get("endpointVersions");
+        Map<String, String> serviceVersion = (Map<String, String>) versions.get("serviceVersion");
+
+        if(serviceVersion != null) {
+            System.out.print("REMReM Publish version ");
+            for (String version: serviceVersion.values()) {
+                System.out.println(version);
+            }
+        }
+        if(endpointVersions != null) {
+            System.out.println("Available endpoints");
+            for (Map.Entry<String, String> entry : endpointVersions.entrySet()) {
+                System.out.println(entry);
+            }
+        }
+        exit(0);
     }
 }