+++ publish-service/src/main/java/com/ericsson/eiffel/remrem/publish/controller/ProducerController.java
@@ -14,6 +14,8 @@
 */
 package com.ericsson.eiffel.remrem.publish.controller;
 
+import java.util.Map;
+
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -29,12 +31,14 @@ import com.ericsson.eiffel.remrem.protocol.MsgService;
 import com.ericsson.eiffel.remrem.publish.helper.PublishUtils;
 import com.ericsson.eiffel.remrem.publish.service.MessageService;
 import com.ericsson.eiffel.remrem.publish.service.SendResult;
+import com.ericsson.eiffel.remrem.shared.VersionService;
 import com.google.gson.JsonElement;
+import com.google.gson.JsonParser;
 
 import ch.qos.logback.classic.Logger;
 
 @RestController
-@RequestMapping("/producer")
+@RequestMapping("/*")
 public class ProducerController {
 
     @Autowired
@@ -45,7 +49,7 @@ public class ProducerController {
     Logger log = (Logger) LoggerFactory.getLogger(ProducerController.class);
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
-    @RequestMapping(value = "/msg", method = RequestMethod.POST)
+    @RequestMapping(value = "/producer/msg", method = RequestMethod.POST)
     @ResponseBody
     public ResponseEntity send(@RequestParam(value = "mp", required = false) String msgProtocol,
             @RequestParam(value = "ud", required = false) String userDomain, @RequestBody JsonElement body) {
@@ -55,5 +59,15 @@ public class ProducerController {
         log.debug("body: " + body);
         SendResult result = messageService.send(body, msgService, userDomain);
         return new ResponseEntity(result, messageService.getHttpStatus());
+    }
+    
+    /**
+     * @return this method returns the current version of publish and all loaded protocols.
+     */
+    @RequestMapping(value = "/versions", method = RequestMethod.GET)
+    public JsonElement getVersions() {
+        JsonParser parser = new JsonParser();
+        Map<String, Map<String, String>> versions = new VersionService().getMessagingVersions();
+        return parser.parse(versions.toString());
     }
 }